requirements.txt
pyspark==3.5.1
requests==2.32.3
paho-mqtt==2.1.0
confluent-kafka==2.4.0
Faker==25.3.0
pyyaml==6.0.1